+++ src/ios/CordovaHttpPlugin.h
@@ -5,7 +5,6 @@
 @interface CordovaHttpPlugin : CDVPlugin
 
 - (void)setServerTrustMode:(CDVInvokedUrlCommand*)command;
-- (void)disableRedirect:(CDVInvokedUrlCommand*)command;
 - (void)post:(CDVInvokedUrlCommand*)command;
 - (void)get:(CDVInvokedUrlCommand*)command;
 - (void)put:(CDVInvokedUrlCommand*)command;